1. Модульные тесты. [#159435 #263329]

+++ src/main/java/ru/job4j/converter/Converter.java
@@ -18,12 +18,12 @@ public class Converter {
         System.out.println("637 rubles are " + euro + " euro.");
         System.out.println("243 rubles are " + dollar + " dollar.");
         float in = 637;
-        float expected = (float) 9.1;
+        float expected = 9.1f;
         float out = Converter.rubleToEuro(in);
         boolean passed = expected == out;
         System.out.println("637 rubles are 9.1 Test result : " + passed);
         float in2 = 243;
-        float expected2 = (float) 4.05;
+        float expected2 = 4.05f;
         float out2 = Converter.rubleToDollar(in2);
         boolean passed2 = expected2 == out2;
         System.out.println("243 rubles are 4.05 Test result : " + passed2);
